FINISHED delete / modify | ADDED YourPost page | sweetalert with react

+++ backend/controllers/postsCtrl.js
@@ -2,7 +2,7 @@ const post = require('../models/postsModel');
 const user = require('../models/usersModel');
 const fs = require('fs');
 
-//Read All and Read One 
+//Read All, One, User's posts 
 exports.getAllPosts = (req, res, next) => {
     post.find().sort({ date: 'descending' })
         .then(posts => res.status(200).json(posts))
@@ -15,20 +15,15 @@ exports.getOnePost = (req, res, next) => {
         .catch(error => res.status(400).json({ error }));
 };
 
-// //Read Liked
-// exports.getLikedPosts = (req, res, next) => {
-//     req.liked.forEach(_id => {
-//         post.find({ _id: req.params.id })
-//             .then()// ici il faudra utiliser un accumulateur avant de faire un res finale quand le nouveau tableau sera rempli
-//             // new array => populate array => ".then(newarray=> res.status(200).json(newarray)"
-//             .catch(error => res.status(500).json({ error }));
-//     })
-// };
+exports.getUserPosts = (req, res, next) => {
+    post.find({ authorId: req.auth.userId }).sort({ date: 'descending' })
+        .then(posts => res.status(200).json(posts))
+        .catch(error => res.status(400).json({ error: "user not found or unidentified" }));
+};
 
 //Create
 exports.createPost = (req, res, next) => {
     const postObject = req.body
-    // console.log('postCtrl Log : ', req.body)
     delete postObject._id;
     const newPost = new post({
         ...postObject,
@@ -44,14 +39,25 @@ exports.createPost = (req, res, next) => {
 //Update
 exports.modifyPost = (req, res, next) => {
     let postObject = req.body
+    console.log(req.file)
     post.findOne({ _id: req.params.id })
         .then(oldPost => {
             if (req.file) {
-                let linkImgToRemove = oldPost.imageUrl.split(`${req.protocol}://${correctURL}`)
-                fs.unlink('./' + linkImgToRemove[1], (error) => {
-                    if (error) {
-                        console.log(error)
-                    }
+                if (oldPost.imageUrl) {
+                    let linkImgToRemove = oldPost.imageUrl.split(`${req.protocol}://localhost:3001/`)
+                    fs.unlink('./' + linkImgToRemove[1], (error) => {
+                        if (error) {
+                            console.log(error)
+                        }
+                        postObject = {
+                            ...postObject,
+                            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+                        }
+                        post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
+                            .then(() => res.status(200).json({ message: 'publication mise à jour' }))
+                            .catch(error => res.status(400).json({ error }));
+                    })
+                } else {
                     postObject = {
                         ...postObject,
                         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
@@ -59,7 +65,7 @@ exports.modifyPost = (req, res, next) => {
                     post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
                         .then(() => res.status(200).json({ message: 'publication mise à jour' }))
                         .catch(error => res.status(400).json({ error }));
-                })
+                }
             } else {
                 post.updateOne({ _id: req.params.id }, { ...postObject, _id: req.params.id })
                     .then(() => res.status(200).json({ message: 'publication mise à jour' }))
@@ -71,78 +77,86 @@ exports.modifyPost = (req, res, next) => {
 
 //Delete
 exports.deletePost = (req, res, next) => {
-    post.findOne({ _id: req.params.id })
-        .then((post) => {
-            if (!post) {
-                return res.status(404).json({ message: 'post not found' })
-            }
-            if (post.authorId !== req.auth.userId && user.isAdmin !== true) {
-                return res.status(401).json({ message: 'You cannot suppress this post!' })
-            }
-            const filename = post.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                post.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: 'Publication supprimée' }))
-                    .catch(error => res.status(400).json({ error }));
-            })
+    user.findOne({ _id: req.auth.userId })
+        .then(currentUser => {
+            post.findOne({ _id: req.params.id })
+                .then((post) => { // refactor this in switch ?
+                    if (!post) {
+                        return res.status(404).json({ message: 'post not found' })
+                    }
+                    if (post.authorId !== req.auth.userId && currentUser.isAdmin !== true) {
+                        return res.status(401).json({ message: 'You cannot suppress this post!' })
+                    }
+                    if (!post.imageUrl) {
+                        post.deleteOne({ _id: req.params.id })
+                            .then(() => res.status(200).json({ message: 'Publication supprimée' }))
+                            .catch(error => res.status(400).json({ message: error }));
+                    } else {
+                        const filename = post.imageUrl.split('/images/')[1];
+                        fs.unlink(`images/${filename}`, () => {
+                            post.deleteOne({ _id: req.params.id })
+                                .then(() => res.status(200).json({ message: 'Publication supprimée' }))
+                                .catch(error => res.status(400).json({ message: error }));
+                        })
+                    }
+                })
+                .catch(error => res.status(400).json({ message: error.Error }))
+
         })
-        .catch(error => res.status(400).json({ message: error.Error }));
+        .catch(error => res.status(401).json({ error: "You dont have credentials" }))
+
 };
 
 
 //likes and dislikes
 exports.likes = (req, res, next) => {
-    console.log(req.auth)
-    user.findOne({ _id: req.auth.userId })
-        .then((currentUser) => {
-            console.log("userfound", current._id, 'now need the post', req.params.id)
 
-            post.findOne({ _id: req.params.id })
+    const currentUserId = req.auth.userId
+    const currentPostId = req.params.id
+    user.findById(currentUserId)
+        .then((currentUser) => {
+            post.findById(currentPostId)
                 .then((currentPost) => {
+                    console.log(currentPost._id)
+                    console.log(currentUser.userLiked.includes(currentPostId))
+
                     //check user like status => if user already likes this do nothing more, else increment likes.
                     if (req.body.like == 1) {
-                        if (currentUser.userLiked.include(post._id)) {
-                            console.log("you already liked this");
+                        // if users already liked, cancels the like
+                        if (currentUser.userLiked.includes(currentPostId)) {
+                            // console.log("you already liked this");
+                            currentPost.likes--
+                            const thisPostIndex = currentUser.userLiked.indexOf(currentPost._id)
+                            currentUser.userLiked.splice(thisPostIndex, 1)
                         } else {
                             currentPost.likes++
-                            currentUser.userLiked.push(post._id)
+                            currentUser.userLiked.push(currentPostId)
                         }
                     }
 
                     //check user dislike status => if user already dislikes this do nothing more, else increment dislikes.
                     if (req.body.like == -1) {
-                        if (currentUser.userDisliked.include(post._id)) {
-                            console.log("you already disliked this");
+                        // if users already disliked, cancels the dislike
+                        if (currentUser.userDisliked.includes(currentPostId)) {
+                            // console.log("you already disliked this");
+                            currentPost.dislikes--
+                            const thisPostIndex = currentUser.userDisliked.indexOf(currentPost._id)
+                            currentUser.userDisliked.splice(thisPostIndex, 1)
+
                         } else {
                             currentPost.dislikes++
-                            currentUser.userDisliked.push(post._id)
+                            currentUser.userDisliked.push(currentPostId)
                         }
                     }
-
-                    // cancels likes or dislikes
-                    else if (req.body.like == 0) {
-                        if (currentUser.userLiked.includes(post._id)) {
-                            const thisPostIndex = currentUser.userLiked.indexOf(post._id)
-                            currentUser.userLiked.splice(thisPostIndex, 1)
-                            currentPost.likes--
-                        }
-                        if (currentUser.usersDisliked.includes(post._id)) {
-                            const thisPostIndex = currentUser.userDisliked.indexOf(post._id)
-                            currentUser.userDisliked.splice(thisPostIndex, 1)
-                            currentPost.dislikes--
-                        }
-
-                    }
-                    // currentPost.likes = post.likes.length
-                    // currentPost.dislikes = post.dislikes.length
-                    post.save()
-                        .then((post) => res.status(200).json({ message: 'likes et dislikes mis à jour' }))
+                    currentPost.save()
+                        .then((post) => {
+                            currentUser.save()
+                                .then(() => res.status(200).json({ message: 'likes et dislikes mis à jour' }))
+                                .catch(() => res.status(400).json({ "error in save ": new Error }));
+                        })
                         .catch(() => res.status(400).json({ "error in save ": new Error }));
                 })
                 .catch(error => res.status(400).json({ "error in post findOne ": error }));
-
         })
         .catch(error => res.status(400).json({ "error in userfindOne ": error }));
-
-
 };
